% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adls_timevarying_region_data}
\alias{adls_timevarying_region_data}
\title{Simulated Time-Varying Residence Data}
\format{
\subsection{\code{andh_forest_data}}{

A data frame with 546 rows and 7 columns describing 100 people:
\describe{
\item{id}{todo}
\item{dob}{todo}
\item{region}{todo}
\item{move_in}{todo}
\item{move_out}{todo}
\item{gender}{todo}
\item{claim}{todo}
}
}
}
\usage{
adls_timevarying_region_data
}
\description{
A dataset of simulated time-varying recidence and gender data.
}
\keyword{datasets}
