% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{andh_forest_data}
\alias{andh_forest_data}
\title{Example Data for Husby's Forest Plot Vignette}
\format{
\subsection{\code{andh_forest_data}}{

A data frame with 18 rows and 12 columns:
\describe{
\item{type}{todo}
\item{indent}{todo}
\item{text}{todo}
\item{A_est}{todo}
\item{A_l}{todo}
\item{A_u}{todo}
\item{B_est}{todo}
\item{B_l}{todo}
\item{B_u}{todo}
\item{C_est}{todo}
\item{C_l}{todo}
\item{C_u}{todo}
}
}
}
\usage{
andh_forest_data
}
\description{
A data example for the construction of a multi faceted forest plot.
}
\keyword{datasets}
