% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_covariate_balance.R
\name{ceiling_dec}
\alias{ceiling_dec}
\title{Round numbers up to a given number of decimal places}
\usage{
ceiling_dec(x, digits = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\value{
The rounded up numeric vector
}
\description{
Round numbers up to a given number of decimal places
}
