% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_fct_confint.R
\name{ci_fct_error_handler}
\alias{ci_fct_error_handler}
\title{Handle errors returned by ci_fct}
\usage{
ci_fct_error_handler(e, which_parm, env)
}
\arguments{
\item{e}{error returned by ci_fct}

\item{which_parm}{Either a logical vector the same length as the coefficient
vector, with \code{TRUE} indicating a coefficient is used by \code{f}, or an integer
vector with the indices of the coefficients used by \code{f}.}

\item{env}{environment to assign n_grid and k}
}
\value{
returns NULL if no stop command is executed.
}
\description{
Handle errors returned by ci_fct
}
\examples{
1+1
}
