% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_covariate_balance.R
\name{CovariateBalance}
\alias{CovariateBalance}
\title{Plots for checking covariate balance in causal forest}
\usage{
CovariateBalance(
  cf,
  covariates = dplyr::everything(),
  names = NULL,
  treatment_name = "W",
  asmd_breaks = NULL,
  asmd_xlim = NULL,
  asmd_scale_color = NULL,
  cd_nrow = NULL,
  cd_ncol = NULL,
  cd_x_scale_width = NULL,
  cd_bar_width = NULL,
  cd_scale_fill = NULL
)
}
\arguments{
\item{cf}{An object of class causal_forest (and inheriting from class grf).}

\item{covariates}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> An unquoted expression
selecting covariates to include in the balance plots.}

\item{names}{A named character vector. The vector itself should contain
covariate names from the causal_forest object, while the names attribute
should contain the names to use when plotting. If the vector is unnamed,
the provided vector will act as the new covariate names, given in the order
of \code{cf$X_orig}. If \code{NULL} (the default), the original names are used.}

\item{treatment_name}{Character, name of treatment.}

\item{asmd_breaks}{Numeric, breaks used in the plot of absolute standardized
mean differences.}

\item{asmd_xlim}{Numeric, \code{x}-limits used in the plot of absolute
standardized mean differences.}

\item{asmd_scale_color}{Function, \code{scale_color_.} function to use in the plot
of absolute standardized mean differences.}

\item{cd_nrow, cd_ncol}{Numeric, the dimensions of the grid to create in
covariate distribution plots. If both are \code{NULL} it will use the same logic
as \link[ggplot2]{facet_wrap} to set the dimensions.}

\item{cd_x_scale_width}{Numeric, the distance between major \code{x}-axis tics in
the covariate distribution plots. If \code{NULL}, a width is chosen to display
approximately six major tics. If length 1, the same width is used for all
covariate plots. If the same length as the number of covariates included,
each number is used as the width for different covariates, in the order of
the covariates after selection with the tidy-select expression in
\code{covariates}.}

\item{cd_bar_width}{Numeric, the width of the bars in the covariate
distribution plots (barplots for categorical variables, histograms for
continuous variables). If \code{NULL}, a width is chosen to display
approximately 50 bars in histograms, while 0.9 times the resolution of the
data is used in bar plots. If length 1, the same width is used for all
covariate plots. This is not recommended if there are both categorical and
continuous covariates. If the same length as the number of covariates
included, each number is used as the bar width for different covariates, in
the order of the covariates after selection with the tidy-select expression
in \code{covariates}.}

\item{cd_scale_fill}{Function, \code{scale_fill_.} function to use in covariate
distribution plots.}
}
\value{
A list with five elements:
\itemize{
\item asmd_data: data used to plot the absolute standardized mean differences.
\item asmd: plot object for absolute standardized mean differences.
\item cd_data: data used to plot covariate distributions.
\item cd_unadjusted: plot of unadjusted covariate distributions in the exposure
groups.
\item cd_adjusted: plot of adjusted covariate distributions in the exposure
groups.
}
}
\description{
Generate plots showing balance in the covariates before and after propensity
score weighting with a causal forest object.
}
\details{
The plot theme can be adjusted using ggplot2 active theme modifiers,
see \link[ggplot2]{theme_get}.
}
\examples{
n <- 1500
p <- 5
X <- matrix(rnorm(n * p), n, p) |> as.data.frame()
expo_prob <- 1 / (1 + exp(0.2 * X[, 1] + 0.1 * X[, 2] - 0.3 * X[, 3]))
W <- rbinom(n, 1, expo_prob)
event_prob <- 1 / (1 + exp(2 * (pmax(2 * X[, 1], 0) * W - X[, 2])))
Y <- rbinom(n, 1, event_prob)
cf <- grf::causal_forest(X, Y, W)
cb1 <- CovariateBalance(cf)
cb2 <- CovariateBalance(
cf,
covariates = 1:4,
names = c(
"medium imbalance" = "V1",
"low imbalance" = "V2",
"high imbalance" = "V3",
"no imbalance" = "V4"
),
treatment_name = "Treatment",
asmd_breaks = seq(0, 0.5, 0.1),
asmd_xlim = c(0, 0.5),
cd_nrow = 2,
cd_x_scale_width = 1,
cd_bar_width = 0.3
)

}
\author{
KIJA
}
