% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_covariate_balance.R
\name{decimalplaces}
\alias{decimalplaces}
\title{Determine number of decimal places}
\usage{
decimalplaces(x)
}
\arguments{
\item{x}{Numeric, a single decimal number}
}
\value{
The number of decimal places in x
}
\description{
Determine number of decimal places
}
