% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiForsk_package.R
\docType{data}
\name{.datatable.aware}
\alias{.datatable.aware}
\title{make package data table aware}
\format{
An object of class \code{logical} of length 1.
}
\usage{
.datatable.aware
}
\description{
This package uses data.table as a fast alternative to dplyr in cases where
performance is essential.
}
\keyword{datasets}
