% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_fct_confint.R
\name{fct_confint}
\alias{fct_confint}
\title{Confidence set for functions of model parameters}
\usage{
fct_confint(
  object,
  f,
  which_parm = rep(TRUE, length(coef(object))),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{A fitted model object.}

\item{f}{A function taking the parameter vector as its single argument, and
returning a numeric vector.}

\item{which_parm}{Either a logical vector the same length as the coefficient
vector, with TRUE indicating a coefficient is used by f, or an integer vector
with the indices of the coefficients used by f.}

\item{level}{The confidence level required.}

\item{...}{Additional argument(s) passed to methods.}
}
\value{
A tibble with columns estimate, conf.low, and conf.high.
}
\description{
TODO
}
\details{
TODO
}
\section{Author(s)}{
KIJA
}

\examples{
1+1

}
