% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_covariate_balance.R
\name{floor_dec}
\alias{floor_dec}
\title{Round numbers down to a given number of decimal places}
\usage{
floor_dec(x, digits = 1)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{integer indicating the number of decimal places}
}
\value{
The rounded down numeric vector
}
\description{
Round numbers down to a given number of decimal places
}
