% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_lms.R
\name{lms}
\alias{lms}
\title{Wrapper around lm for sibling design}
\usage{
lms(data, formula, grp_id, obs_id = NULL, ...)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{formula}{A formula, used to create the model matrix, whose columns are
demeaned.}

\item{grp_id}{<\code{\link[=dplyr_data_masking]{data-masking}}> One unquoted expression
naming the id variable in data defining the groups to demean,
e.g. sibling groups.}

\item{obs_id}{<\code{\link[=dplyr_data_masking]{data-masking}}> Optional, One unquoted
expression naming an id variable to keep track of the input data order.}

\item{...}{Additional arguments to be passed to \link[stats]{lm}().}
}
\value{
A list with class (lms, lm). Contains the output from lm applied to
demeaned data according to formula and the original data and provided
formula.
}
\description{
Fits a linear model using demeaned data. Useful for sibling design.
}
\details{
TODO
}
\section{Author(s)}{
KIJA
}

\examples{

1+1

}
