% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aso_many_merge.R
\name{many_merge}
\alias{many_merge}
\title{Merging Many Data Frames with Name Handling}
\usage{
many_merge(by, first_data, ...)
}
\arguments{
\item{by}{A join specification created with
\code{\link[dplyr:join_by]{join_by()}}, or a character vector of variables
to join by. The \code{by} must be present in all data frames \code{first_data} and
\code{...}.}

\item{first_data}{A data frame (presented on the left in the final table).}

\item{...}{Data frames to merge onto \code{first_data}.}
}
\value{
The \code{many_merge()} function returns a data frame.
}
\description{
Function to join/merge multiple data.frames with one or more common variable
names.
}
\examples{

# Create some dummy data
testdata_id <- c(1:10)
var1 <- rep(letters[1:5], times = 2)
var2 <- letters[1:10]
var3 <- rep(letters[11:12], times = 5)
var4 <- letters[13:22]
var5 <- letters[11:20]

# Rename alle the variables to "var"
data1 <- data.frame(testdata_id, var = var1)
data2 <- data.frame(testdata_id, var = var2)
data3 <- data.frame(testdata_id, var = var3)
data4 <- data.frame(testdata_id, var = var4)
data5 <- data.frame(testdata_id, var = var5)

# Many merge
final_data <- many_merge(
  by = c("testdata_id"),
  data1,
  data2,
  data3,
  data4,
  data5
)

}
\author{
ASO
}
