% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_print_lms.R
\name{print.lms}
\alias{print.lms}
\title{Print method for lms objects}
\usage{
\method{print}{lms}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{The S3 object with class lms.}

\item{digits}{The number of significant digits to be passed to
\link[base]{format}(\link[stats]{coef}(x), .) when \link[base]{print}()ing.}

\item{...}{Any additional arguments are ignored without warning.}
}
\value{
Silently returns the input lms object.
}
\description{
Prints information about an lms object.
}
\details{
TODO
}
\section{Author(s)}{
KIJA
}

\examples{

1+1

}
