% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aso_odds_ratio_function.R
\name{summary.svy_vglm}
\alias{summary.svy_vglm}
\title{Summary function for svy_vglm objects}
\usage{
\method{summary}{svy_vglm}(object, ...)
}
\arguments{
\item{object}{An svy_vglm object}

\item{...}{additional arguments. Not used.}
}
\value{
A "summary.svy_vglm" object is returned.
}
\description{
Internal summary function for svy_vglm objects
}
