% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kija_vcovHC.R
\name{vcovHC}
\alias{vcovHC}
\title{Heteroscedasticity-Consistent Covariance Matrix}
\usage{
vcovHC(x)
}
\arguments{
\item{x}{lm object}
}
\value{
A matrix containing the covariance matrix estimate.
}
\description{
Calculate Heteroscedasticity-Consistent Covariance Matrix from a linear
model using the HC3 method from sandwich.
}
\examples{
1+1
}
