% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adls_timevarying_region_data}
\alias{adls_timevarying_region_data}
\title{Simulated Time-Varying Residence Data}
\format{
\subsection{\code{andh_forest_data}}{

A data frame with 546 rows and 7 columns describing 100 people:
\describe{
\item{id}{an id number}
\item{dob}{date of birth}
\item{region}{region of Denmark}
\item{move_in}{date of moving to region}
\item{move_out}{date of moving away from region}
\item{gender}{gender of the person}
\item{claim}{whether or not the person made a claim here}
}
}
}
\usage{
adls_timevarying_region_data
}
\description{
A dataset of simulated time-varying residence and gender data.
}
\keyword{datasets}
