% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{andh_forest_data}
\alias{andh_forest_data}
\title{Example Data for Husby's Forest Plot Vignette}
\format{
\subsection{\code{andh_forest_data}}{

A data frame with 18 rows and 12 columns:
\describe{
\item{type}{text formattaing, bold/plain}
\item{indent}{number of indents in final formatting}
\item{text}{description text}
\item{A_est}{point estimate in first figure column}
\item{A_l}{lower limit of confidence interval in first figure column}
\item{A_u}{upper limit of confidence interval in first figure column}
\item{B_est}{point estimate in second figure column}
\item{B_l}{lower limit of confidence interval in second figure column}
\item{B_u}{upper limit of confidence interval in second figure column}
\item{C_est}{point estimate in third figure column}
\item{C_l}{lower limit of confidence interval in third figure column}
\item{C_u}{upper limit of confidence interval in third figure column}
}
}
}
\usage{
andh_forest_data
}
\description{
A data example for the construction of a multi faceted forest plot.
}
\keyword{datasets}
