% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiForsk_package.R
\docType{package}
\name{EpiForsk-package}
\alias{EpiForsk}
\alias{EpiForsk-package}
\title{EpiForsk}
\description{
This is a collection of assorted functions and examples collected
from various projects. Currently we have functionalities for simplifying
overlapping time intervals, Charlson comorbidity score constructors for
Danish data, getting frequency for multiple variables, getting standardized
output from logistic and log-linear regressions, sibling design linear
regression functionalities a method for calculating the confidence intervals
for functions of parameters from a GLM, Bayes equivalent for hypothesis
testing with asymptotic Bayes factor, and several help functions for
generalized random forest analysis using the grf package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Laksafoss/EpiForsk}
  \item Report bugs at \url{https://github.com/Laksafoss/EpiForsk/issues}
}

}
\author{
\strong{Maintainer}: Kim Daniel Jakobsen \email{kija@ssi.dk} (\href{https://orcid.org/0000-0003-0086-9980}{ORCID})

Authors:
\itemize{
  \item Anders Husby \email{andh@ssi.dk} (\href{https://orcid.org/0000-0002-7634-8455}{ORCID})
  \item Anna Laksafoss \email{adls@ssi.dk} (\href{https://orcid.org/0000-0002-9898-2924}{ORCID})
  \item Emilia Myrup Thiesson \email{emth@ssi.dk} (\href{https://orcid.org/0000-0001-6258-4177}{ORCID})
  \item Mikael Andersson \email{aso@ssi.dk} (\href{https://orcid.org/0000-0002-0114-2057}{ORCID})
  \item Klaus Rostgaard \email{klp@ssi.dk} (\href{https://orcid.org/0000-0001-6220-9414}{ORCID})
}

}
