% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiForsk_package.R
\docType{package}
\name{EpiForsk-package}
\alias{EpiForsk}
\alias{EpiForsk-package}
\title{EpiForsk}
\description{
This is a collection of assorted functions and examples collected from
various projects. Currently we have functionalities for simplifying
overlapping time intervals, Charlson comorbidity score constructors for
Danish data, sibling design linear regression functionalities, and a method
for calculating the confidence intervals for functions of parameters from a
GLM.
}
\details{
In the package there are contributions from
\itemize{
\item{ADLS : }{Anna Damkjær Laksafoss (https://orcid.org/0000-0002-9898-2924)}
\item{ANDH : }{Anders Husby (https://orcid.org/0000-0002-7634-8455)}
\item{ASO  : }{Mikael Andersson}
\item{EMTH : }{Emilia Myrup Thiesson}
\item{KIJA : }{Kim Daniel Jakobsen}
}
}
\author{
\strong{Maintainer}: Anna Laksafoss \email{adls@ssi.dk} (\href{https://orcid.org/0000-0002-9898-2924}{ORCID})

}
