% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiForsk_package.R
\docType{package}
\name{EpiForsk-package}
\alias{EpiForsk}
\alias{EpiForsk-package}
\title{EpiForsk}
\description{
A nice short description of the package.
}
\author{
\strong{Maintainer}: Anna Laksafoss \email{adls@ssi.dk} (\href{https://orcid.org/0000-0002-9898-2924}{ORCID})

}
